triposr-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
Pillow==10.2.0
torch==2.2.0
torchvision==0.17.0
trimesh==4.1.4
rembg==2.0.56
huggingface-hub==0.20.3
